Fix asQueryParams URL encoding issues for GET requests
When the field or other type of data has html entities in them (i.e.
ampersands), since the data is not encoded for GET requests the
JSORM url will be incidentally broken.  For obvious reasons, POST
requests shouldn't have this issue.

encodeURIComponent encodes a lot of useful characters that doesn't
actually need to be encoded for the brower, so we are only encoding
the tokens that are likely to cause issues.